client/oidc: Make client options available in test only
